\name{EnsDb.Hsapiens.v75}
\docType{package}

\alias{EnsDb.Hsapiens.v75-package}
\alias{EnsDb.Hsapiens.v75}
\alias{EnsDb.Hsapiens.v75}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Mon Oct  3 09:48:13 2016 and based on the 75
}

\author{Johannes Rainer}



\examples{
## load the library
##library(EnsDb.Hsapiens.v75)
## list the contents that are loaded into memory
ls('package:EnsDb.Hsapiens.v75')
## show the db object that is loaded by calling it's name
EnsDb.Hsapiens.v75

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
